requirements.txt
matplotlib~=3.7.2
numpy~=1.25.2
pandas~=2.0.3
scikit-learn~=1.3.0
scipy~=1.11.2
torch~=2.1.1
wfdb~=4.1.2
lightning~=2.0.7
PyYAML~=6.0.1
tensorboardx==2.6.2.2
mlflow==2.16.0
seaborn==0.13.2
pytest~=7.4.0
python-dotenv~=1.0.0
